refactor(osd): relocate OSD utility files to components/tabs/osd
Move osd.js, osd_constants.js, and osd_positions.js from
src/js/tabs/ to src/components/tabs/osd/, co-locating them with
the Vue tab component. Update all imports in stores, Vue components,
and test files. Removes the now-empty src/js/tabs/ directory.
No functional changes.
