+++ src/components/tabs/osd/osd_positions.js
